chore(components): remove some circular imports
