+++ packages/components/src/components/form/Select/types.ts
@@ -1,0 +1,1 @@
+export type Option = any;
